[8.x] Backport 119578 + 114289 (#119638)
* Refactor change point detection (#114289)

* Move change detection code to separate class

* Uniformize ChangeDetector and SkipeAndDipDetector

* Separate ChangeDetectorTests and ChangePointAggregatorTests.

* Public entrypoint for change point detection

* Move p-value computation to ChangeDetector

* Move main entrypoint to a separate file

* Fix change point detection for uncertain non-stationary distributions. (#119578)

* Fix change point detection for uncertain non-stationary distributions.

* Replace -1 by ChangeType.NO_CHANGE_POINT
