+++ x-pack/plugin/ml/src/main/java/org/elasticsearch/xpack/ml/aggs/changepoint/SpikeAndDipDetector.java
@@ -92,6 +92,7 @@ final class SpikeAndDipDetector {
         return newValues;
     }
 
+    private final MlAggsHelper.DoubleBucketValues bucketValues;
     private final int numValues;
     private final int dipIndex;
     private final int spikeIndex;
@@ -100,7 +101,9 @@ final class SpikeAndDipDetector {
     private final KDE spikeTestKDE;
     private final KDE dipTestKDE;
 
-    SpikeAndDipDetector(double[] values) {
+    SpikeAndDipDetector(MlAggsHelper.DoubleBucketValues bucketValues) {
+        this.bucketValues = bucketValues;
+        double[] values = bucketValues.getValues();
 
         numValues = values.length;
 
@@ -135,7 +138,7 @@ final class SpikeAndDipDetector {
         spikeTestKDE = new KDE(spikeKDEValues, 1.36);
     }
 
-    ChangeType at(double pValueThreshold, MlAggsHelper.DoubleBucketValues bucketValues) {
+    ChangeType detect(double pValueThreshold) {
         if (dipIndex == -1 || spikeIndex == -1) {
             return new ChangeType.Indeterminable(
                 "not enough buckets to check for dip or spike. Requires at least [3]; found [" + numValues + "]"
