+++ x-pack/plugin/ml/src/test/java/org/elasticsearch/xpack/ml/aggs/changepoint/SpikeAndDipDetectorTests.java
@@ -25,14 +25,14 @@ public class SpikeAndDipDetectorTests extends ESTestCase {
             Arrays.fill(docCounts, 1);
             Arrays.fill(values, 1.0);
             MlAggsHelper.DoubleBucketValues bucketValues = new MlAggsHelper.DoubleBucketValues(docCounts, values);
-            SpikeAndDipDetector detect = new SpikeAndDipDetector(values);
-            assertThat(detect.at(0.01, bucketValues), instanceOf(ChangeType.Indeterminable.class));
+            SpikeAndDipDetector detect = new SpikeAndDipDetector(bucketValues);
+            assertThat(detect.detect(0.01), instanceOf(ChangeType.Indeterminable.class));
         }
     }
 
     public void testSpikeAndDipValues() {
         double[] values = new double[] { 2.0, 1.0, 3.0, 5.0, 4.0 };
-        SpikeAndDipDetector detector = new SpikeAndDipDetector(values);
+        SpikeAndDipDetector detector = new SpikeAndDipDetector(new MlAggsHelper.DoubleBucketValues(null, values));
         assertThat(detector.spikeValue(), equalTo(5.0));
         assertThat(detector.dipValue(), equalTo(1.0));
     }
@@ -133,7 +133,7 @@ public class SpikeAndDipDetectorTests extends ESTestCase {
         Arrays.sort(expectedSpikeKDEValues);
         Arrays.sort(expectedDipKDEValues);
 
-        SpikeAndDipDetector detector = new SpikeAndDipDetector(values);
+        SpikeAndDipDetector detector = new SpikeAndDipDetector(new MlAggsHelper.DoubleBucketValues(null, values));
 
         assertThat(detector.spikeValue(), equalTo(10.0));
         assertThat(detector.dipValue(), equalTo(-2.0));
@@ -150,9 +150,9 @@ public class SpikeAndDipDetectorTests extends ESTestCase {
             double[] values = new double[] { 0.1, 3.1, 1.2, 1.7, 0.9, 2.3, -0.8, 3.2, 1.2, 1.3, 1.1, 1.0, 8.5, 0.5, 2.6, 0.7 };
             MlAggsHelper.DoubleBucketValues bucketValues = new MlAggsHelper.DoubleBucketValues(docCounts, values);
 
-            SpikeAndDipDetector detect = new SpikeAndDipDetector(values);
+            SpikeAndDipDetector detect = new SpikeAndDipDetector(bucketValues);
 
-            ChangeType change = detect.at(0.05, bucketValues);
+            ChangeType change = detect.detect(0.05);
 
             assertThat(change, instanceOf(ChangeType.Spike.class));
             assertThat(change.pValue(), closeTo(3.0465e-12, 1e-15));
@@ -162,9 +162,9 @@ public class SpikeAndDipDetectorTests extends ESTestCase {
             double[] values = new double[] { 0.1, 3.1, 1.2, 1.7, 0.9, 2.3, -4.2, 3.2, 1.2, 1.3, 1.1, 1.0, 3.5, 0.5, 2.6, 0.7 };
             MlAggsHelper.DoubleBucketValues bucketValues = new MlAggsHelper.DoubleBucketValues(docCounts, values);
 
-            SpikeAndDipDetector detect = new SpikeAndDipDetector(values);
+            SpikeAndDipDetector detect = new SpikeAndDipDetector(bucketValues);
 
-            ChangeType change = detect.at(0.05, bucketValues);
+            ChangeType change = detect.detect(0.05);
 
             assertThat(change, instanceOf(ChangeType.Dip.class));
             assertThat(change.pValue(), closeTo(1.2589e-08, 1e-11));
@@ -177,9 +177,9 @@ public class SpikeAndDipDetectorTests extends ESTestCase {
         int[] buckets = new int[] { 0, 2, 3, 6, 7, 8, 9, 10, 11, 12, 13, 15, 17, 18, 19, 20 };
         MlAggsHelper.DoubleBucketValues bucketValues = new MlAggsHelper.DoubleBucketValues(docCounts, values, buckets);
 
-        SpikeAndDipDetector detect = new SpikeAndDipDetector(values);
+        SpikeAndDipDetector detect = new SpikeAndDipDetector(bucketValues);
 
-        ChangeType change = detect.at(0.01, bucketValues);
+        ChangeType change = detect.detect(0.01);
 
         assertThat(change, instanceOf(ChangeType.Spike.class));
         assertThat(change.changePoint(), equalTo(10));
